Changing the blog structure to use the author and about contenxt using pages
We're not using the `plugin/author_plus` anymore. Now, only a single
'sobre-o-autor.md' will be used.

The `primeiro.md` was changed to a page (instead an article) too.

The `tags_cloud` plugins was instaled/configured also (but it still not
being used).

Also, a new CSS class is set now in each article/page to be easier to
apply specific CSS rules.